+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/CombatTrackerAccessor.java
@@ -1,0 +1,12 @@
+package insane96mcp.iguanatweaksreborn.mixin;
+
+import net.minecraft.world.damagesource.CombatTracker;
+import net.minecraft.world.entity.LivingEntity;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(CombatTracker.class)
+public interface CombatTrackerAccessor {
+	@Accessor
+	LivingEntity getMob();
+}