+++ src/main/java/insane96mcp/iguanatweaksreborn/module/movement/NoPillaring.java
@@ -7,8 +7,7 @@ import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
-import net.minecraft.world.entity.ai.targeting.TargetingConditions;
-import net.minecraft.world.entity.monster.Monster;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.context.BlockPlaceContext;
@@ -19,19 +18,16 @@ import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.eventbus.api.Event;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-@LoadFeature(module = Modules.Ids.MOVEMENT, description = "Prevents the player from placing blocks below him when in mid air.")
+@LoadFeature(module = Modules.Ids.MOVEMENT, enabledByDefault = false, description = "Prevents the player from placing blocks below him when in mid air.")
 public class NoPillaring extends Feature {
 
 	public static final String NO_PILLARING_LANG = "iguanatweaksreborn.no_pillaring";
 
-	@Config(description = "If true, pillaring will be negated only if there are monsters nearby")
-	public static Boolean monstersNearby = true;
+	@Config(description = "If true, pillaring will be negated only if the last mobs that hit you are nearby")
+	public static Boolean engagedMobs = true;
 
 	@Config(description = "Range at which monsters must be in order to negate pillaring")
-	public static Integer monstersRange = 12;
-
-	@Config(description = "If true, pillaring will be negated only if the player is engaged in combat")
-	public static Boolean engaged = true;
+	public static Integer monstersRange = 24;
 
 	public NoPillaring(Module module, boolean enabledByDefault, boolean canBeDisabled) {
 		super(module, enabledByDefault, canBeDisabled);
@@ -41,22 +37,21 @@ public class NoPillaring extends Feature {
 	public void playerRightClickBlock(PlayerInteractEvent.RightClickBlock event) {
 		if (!this.isEnabled())
 			return;
-		Player playerEntity = event.getEntity();
-		if (playerEntity.isCreative()
-				|| playerEntity.isInWater()
-				|| playerEntity.onClimbable())
+		Player player = event.getEntity();
+		if (player.isCreative()
+				|| player.isInWater()
+				|| player.onClimbable())
 			return;
-        if (monstersNearby
-				&& event.getLevel().getNearestEntity(Monster.class, TargetingConditions.forNonCombat().ignoreLineOfSight(), event.getEntity(), event.getEntity().getX(), event.getEntity().getY(), event.getEntity().getZ(), event.getEntity().getBoundingBox().inflate(monstersRange)) == null)
+        if (engagedMobs && !isAnyCombatEntryMobNearby(player))
             return;
 		//noinspection ConstantConditions
 		BlockPos placedPos = event.getPos().relative(event.getFace());
 		Vec3 placedBlock = new Vec3(placedPos.getX() + 0.5d, placedPos.getY() + 0.5d, placedPos.getZ() + 0.5d);
-		double distance = placedBlock.distanceTo(playerEntity.position());
+		double distance = placedBlock.distanceTo(player.position());
 		double allowedDistance = 1.35d;
-		//if (playerEntity.hasEffect(MobEffects.JUMP))
+		//if (player.hasEffect(MobEffects.JUMP))
 			//noinspection ConstantConditions
-			//allowedDistance *= 1 + ((playerEntity.getEffect(MobEffects.JUMP).getAmplifier() + 1) * 0.5);
+			//allowedDistance *= 1 + ((player.getEffect(MobEffects.JUMP).getAmplifier() + 1) * 0.5);
 		boolean isSolidBlock = true;
 		if (event.getItemStack().getItem() instanceof BlockItem) {
 			Block block = ((BlockItem) event.getItemStack().getItem()).getBlock();
@@ -65,10 +60,18 @@ public class NoPillaring extends Feature {
 				state = block.defaultBlockState();
 			isSolidBlock = state.blocksMotion();/*state.canOcclude();/*state.entityCanStandOn(event.getLevel(), event.getPos(), event.getEntity());*/
 		}
-		if (isSolidBlock && playerEntity.getViewXRot(1.0f) > 40f && !playerEntity.onGround() && event.getItemStack().getItem() instanceof BlockItem && distance <= allowedDistance && playerEntity.getY() > placedPos.getY()) {
+		if (isSolidBlock && player.getViewXRot(1.0f) > 40f && !player.onGround() && event.getItemStack().getItem() instanceof BlockItem && distance <= allowedDistance && player.getY() > placedPos.getY()) {
 			event.setCanceled(true);
 			event.setResult(Event.Result.DENY);
 			event.getEntity().displayClientMessage(Component.translatable(NO_PILLARING_LANG), true);
 		}
+	}
+
+	public static boolean isAnyCombatEntryMobNearby(Player player) {
+		return player.getCombatTracker().entries
+				.stream()
+				.anyMatch(entry
+						-> entry.source().getEntity() instanceof LivingEntity attacker
+						&& player.distanceToSqr(attacker) <= monstersRange * monstersRange);
 	}
 }